+++ knf/validators/fs/validators_test.go
@@ -13,7 +13,7 @@ import (
 	"os"
 	"testing"
 
-	"pkg.re/essentialkaos/ek.v11/knf"
+	"pkg.re/essentialkaos/ek.v12/knf"
 
 	. "pkg.re/check.v1"
 )