+++ path/path_test.go
@@ -8,7 +8,7 @@ package path
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
-	"pkg.re/essentialkaos/ek.v11/env"
+	"pkg.re/essentialkaos/ek.v12/env"
 	"testing"
 
 	. "pkg.re/check.v1"
@@ -43,6 +43,16 @@ func (s *PathUtilSuite) TestBase(c *C) {
 
 	c.Assert(d, Equals, "/some/test/path/")
 	c.Assert(f, Equals, "file.jpg")
+}
+
+func (s *PathUtilSuite) TestDirN(c *C) {
+	c.Assert(DirN("", 99), Equals, "")
+	c.Assert(DirN("1", 99), Equals, "1")
+	c.Assert(DirN("abcde", 1), Equals, "abcde")
+	c.Assert(DirN("/a/b/c/d", -1), Equals, "/a/b/c/d")
+	c.Assert(DirN("/a/b/c/d", 1), Equals, "/a")
+	c.Assert(DirN("a/b/c/d", 2), Equals, "a/b")
+	c.Assert(DirN("/a/b/c/d", 99), Equals, "/a/b/c/d")
 }
 
 func (s *PathUtilSuite) TestEvalHome(c *C) {