+++ pid/pid.go
@@ -16,7 +16,7 @@ import (
 	"strconv"
 	"strings"
 
-	"pkg.re/essentialkaos/ek.v11/fsutil"
+	"pkg.re/essentialkaos/ek.v12/fsutil"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //