+++ system/info_memory_linux.go
@@ -13,7 +13,7 @@ import (
 	"io"
 	"os"
 
-	"pkg.re/essentialkaos/ek.v11/strutil"
+	"pkg.re/essentialkaos/ek.v12/strutil"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //