+++ usage/completion/fish/examples_test.go
@@ -10,7 +10,7 @@ package fish
 import (
 	"fmt"
 
-	"pkg.re/essentialkaos/ek.v11/usage"
+	"pkg.re/essentialkaos/ek.v12/usage"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //