+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 ## Changelog
 
+### 12.0.0
+
+* `[path]` Added method `DirN` for reading N elements from path
+* `[pluralize]` Methods `Pluralize` and `PluralizeSpecial` now return only pluralized word (_incompatible changes_)
+* `[pluralize]` Added methods `P` and `PS` for pluralization with custom formatting
+* `[pluralize]` Added usage examples
+
+---
+
 ### 11.6.3
 
-* `[usage]` Add more examples
+* `[usage]` Added more examples
 
 ### 11.6.2